Re-integrate GUI to main loop, add switch debounce handling.
Further refining/debugging the transmitter tft GUI.

Last presses of buttons are now tracked; if the current press is within
too short of an interval it is ignored. Note that the raw input is still
being transmitted to the car; this is for the tx unit only right now.

The separate tft class was proving to be a bit of a pain so I dropped it
for now.