Add axis-centering function.
Due to the transimitter's construction, the "centered" position
of the input potentiometers (defined either by spring tention
for steering and throttle or arrow direction for the trim pots)
might not correspond to the middle point between minimum and
maximum analog readings.

A centering option has been added to the pause menu screen, and
methods have been added to the Axis class to both track and
reassign the centered value.

The input-norming axis methods have been tweaked to allow for this
change, with different cases for values below and above the center.

The end result is a input center that corresponds to the physical
center, and output maximums and minumums that correspond to the
physical limits of travel for the involved potentiometers.

A side effect is a relative loss of precision for the side of center
that is "shorter", but this should be minor.